Refactor settings page
